move nanoid & close to own file
