+++ src/components/Window.svelte
@@ -1,5 +1,6 @@
 <script lang="ts">
-  import { minimized, open_apps, close } from "../stores";
+  import { minimized, open_apps } from "../stores";
+  import {close} from '../util'
   import type { Session, MouseEventHandler } from "../types";
   import { slide } from "svelte/transition";
   export let session: Session;
